requirements.txt
psycopg2-binary>=2.9.0
sqlalchemy>=1.4.0
pandas>=1.3.0
python-dotenv>=0.19.0
flask>=2.0.0
flask-cors>=3.0.0